+++ sshd-contrib/src/main/java/org/apache/sshd/server/scp/SimpleAccessControlScpEventListener.java
@@ -26,6 +26,7 @@ import java.nio.file.attribute.PosixFilePermission;
 import java.util.Set;
 
 import org.apache.sshd.common.scp.AbstractScpTransferEventListenerAdapter;
+import org.apache.sshd.common.session.Session;
 
 /**
  * Provides a simple access control by making a distinction between methods
@@ -37,12 +38,12 @@ public abstract class SimpleAccessControlScpEventListener extends AbstractScpTra
     public static final SimpleAccessControlScpEventListener READ_ONLY_ACCESSOR =
         new SimpleAccessControlScpEventListener() {
             @Override
-            protected boolean isFileUploadAllowed(Path path) throws IOException {
+            protected boolean isFileUploadAllowed(Session session, Path path) throws IOException {
                 return false;
             }
 
             @Override
-            protected boolean isFileDownloadAllowed(Path path) throws IOException {
+            protected boolean isFileDownloadAllowed(Session session, Path path) throws IOException {
                 return true;
             }
     };
@@ -52,18 +53,19 @@ public abstract class SimpleAccessControlScpEventListener extends AbstractScpTra
     }
 
     @Override
-    public void startFileEvent(FileOperation op, Path file, long length, Set<PosixFilePermission> perms)
-            throws IOException {
-        super.startFileEvent(op, file, length, perms);
+    public void startFileEvent(
+            Session session, FileOperation op, Path file, long length, Set<PosixFilePermission> perms)
+                throws IOException {
+        super.startFileEvent(session, op, file, length, perms);
         switch(op) {
             case SEND:
-                if (!isFileDownloadAllowed(file)) {
+                if (!isFileDownloadAllowed(session, file)) {
                     throw new AccessDeniedException(file.toString());
                 }
                 break;
 
             case RECEIVE:
-                if (!isFileUploadAllowed(file)) {
+                if (!isFileUploadAllowed(session, file)) {
                     throw new AccessDeniedException(file.toString());
                 }
                 break;
@@ -73,17 +75,19 @@ public abstract class SimpleAccessControlScpEventListener extends AbstractScpTra
     }
 
     @Override
-    public void startFolderEvent(FileOperation op, Path file, Set<PosixFilePermission> perms) throws IOException {
-        super.startFolderEvent(op, file, perms);
+    public void startFolderEvent(
+            Session session, FileOperation op, Path file, Set<PosixFilePermission> perms)
+                throws IOException {
+        super.startFolderEvent(session, op, file, perms);
         switch(op) {
             case SEND:
-                if (!isFileDownloadAllowed(file)) {
+                if (!isFileDownloadAllowed(session, file)) {
                     throw new AccessDeniedException(file.toString());
                 }
                 break;
 
             case RECEIVE:
-                if (!isFileUploadAllowed(file)) {
+                if (!isFileUploadAllowed(session, file)) {
                     throw new AccessDeniedException(file.toString());
                 }
                 break;
@@ -93,16 +97,18 @@ public abstract class SimpleAccessControlScpEventListener extends AbstractScpTra
     }
 
     /**
+     * @param session The client/server {@link Session} through which the transfer is being executed
      * @param path The local file/folder path
      * @return {@code true} if client is allowed to read from the specified local path
      * @throws IOException If failed to handle the call
      */
-    protected abstract boolean isFileDownloadAllowed(Path path) throws IOException;
+    protected abstract boolean isFileDownloadAllowed(Session session, Path path) throws IOException;
 
     /**
+     * @param session The client/server {@link Session} through which the transfer is being executed
      * @param path The local file/folder path
      * @return {@code true} if client is allowed to write to the specified local path
      * @throws IOException If failed to handle the call
      */
-    protected abstract boolean isFileUploadAllowed(Path path) throws IOException;
+    protected abstract boolean isFileUploadAllowed(Session session, Path path) throws IOException;
 }